requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
django-bulk-sync==3.2.0
redis==4.5.1
celery==5.2.7
requests==2.28.2
psycopg2==2.9.5
pytest==7.4.0
django-environ==0.9.0